azure-functions/jobassistai-document-processing/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-storage-blob>=12.19.0
azure-ai-documentintelligence>=1.0.0
azure-cosmos>=4.6.0
azure-functions>=1.18.0
python-dotenv>=1.0.0